Improve sidebar metric display
Improve formatting and layout
Display metric params, when possible
