+++ src/inspect_ai/_view/www/src/workspace/sidebar/SidebarScoresView.tsx
@@ -2,6 +2,7 @@ import clsx from "clsx";
 import { Fragment } from "react";
 import { Scores } from "../../types/log";
 import { formatPrettyDecimal } from "../../utils/format";
+import { metricDisplayName } from "../utils";
 import styles from "./SidebarScoresView.module.css";
 
 interface SidebarScoresProps {
@@ -9,6 +10,7 @@ interface SidebarScoresProps {
 }
 
 export const SidebarScoresView: React.FC<SidebarScoresProps> = ({ scores }) => {
+  const showReducer = scores.findIndex((score) => !!score.reducer) !== -1;
   return (
     <div className={styles.container}>
       {scores.map((score) => {
@@ -19,16 +21,23 @@ export const SidebarScoresView: React.FC<SidebarScoresProps> = ({ scores }) => {
             <div
               className={clsx(
                 "text-style-secondary",
-                "text-label",
+                "text-style-label",
                 "text-size-small",
                 styles.metricName,
               )}
             >
               {name}
             </div>
-            {reducer ? (
-              <div className={clsx("text-size-small", styles.metricReducer)}>
-                {reducer}
+            {showReducer ? (
+              <div
+                className={clsx(
+                  "text-size-small",
+                  "text-style-label",
+                  "text-style-secondary",
+                  styles.metricReducer,
+                )}
+              >
+                {reducer || "default"}
               </div>
             ) : (
               ""
@@ -38,14 +47,7 @@ export const SidebarScoresView: React.FC<SidebarScoresProps> = ({ scores }) => {
                 const metric = score.metrics[key];
                 return (
                   <Fragment key={key}>
-                    <div
-                      className={clsx(
-                        "text-style-secondary",
-                        "text-style-label",
-                      )}
-                    >
-                      {metric.name}
-                    </div>
+                    <div className={clsx()}>{metricDisplayName(metric)}</div>
                     <div className={styles.metricValue}>
                       {formatPrettyDecimal(metric.value)}
                     </div>
